sys: 64 bit compatibility
Changes other than DEBUG/printf formatting or different types for void pointer casting
* vfs.h: Increased default vfs buffer sizes for 64 bit
* bloom.h: Fixed typedef for the hashfp_t function pointer
* bytes.h: Check if socklen_t is already defined
* priority_queue_t: Replaced `-1U` literal with PRIORITY_QUEUE_DATA_SIGNALING define

+++ sys/include/bloom.h
@@ -134,7 +134,7 @@ extern "C" {
 /**
  * @brief hash function to use in thee filter
  */
-typedef uint32_t (*hashfp_t)(const uint8_t *, int len);
+typedef uint32_t (*hashfp_t)(const uint8_t *, size_t len);
 
 /**
  * @brief bloom_t bloom filter object
